Use (void) in prototypes instead of ().
Also, removed `__inline` from an extern function (why was it there?).
Also, moved one extern declaration to a header file. There are lots of
these that need to be moved, but one thing at a time.

+++ src/chat.c
@@ -936,7 +936,7 @@ static void draw_infobox(ToxWindow *self)
     wattron(infobox->win, A_BOLD);
     wprintw(infobox->win, " VAD level: ");
     wattroff(infobox->win, A_BOLD);
-    wprintw(infobox->win, "%.2f\n", infobox->vad_lvl);
+    wprintw(infobox->win, "%.2f\n", (double) infobox->vad_lvl);
 
     wborder(infobox->win, ACS_VLINE, ' ', ACS_HLINE, ACS_HLINE, ACS_TTEE, ' ', ACS_LLCORNER, ' ');
     wnoutrefresh(infobox->win);
